aarch64: Fix reg_is_wrapped_separately array size [PR100211]
Currrently the size of the array reg_is_wrapped_separately is LAST_SAVED_REGNUM.
But LAST_SAVED_REGNUM could be regno that is being saved. So the size needs
to be `LAST_SAVED_REGNUM + 1` like aarch64_frame->reg_offset is.

Committed as obvious after a bootstrap/test for aarch64-linux-gnu.

gcc/ChangeLog:

	PR target/100211
	* config/aarch64/aarch64.h (machine_function): Fix the size
	of reg_is_wrapped_separately.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>